Removed support for macOS as support is very bad
We have added support for macOS back in March/April 2021 when we had more than enough resources to actually test this library in all the three operating systems.

Because of environmental changes that recently happened, we no longer have enough resources to continue development on macOS, so we removed all support for macOS.

We apologize for your inconvenience.